requirements.txt
torch>=2.0.0
numpy>=1.23.0
opencv-python>=4.5.0
Pillow>=9.0.0 